+++ src/dynl_unix.cpp
@@ -4,41 +4,20 @@
 
 #include <dynl/dynl.hpp>
 
-#include <dlfcn.h>
+#include <dynl/platform_details/dynl_headeronly.hpp>
 
 namespace dynl::_backend {
-namespace {
-dynamic_library_pointer *as_placeholder(void *in) {
-  return reinterpret_cast<dynamic_library_pointer *>(in);
-}
-void *as_native_lib(dynamic_library_pointer *in) {
-  return reinterpret_cast<dynamic_library_pointer *>(in);
-}
-} // namespace
 
-dynamic_function_symbol do_find_function(dynamic_library_pointer *lib,
+dynamic_function_symbol src_find_function(dynamic_library_pointer *lib,
                                          char const *name,
                                          error_callback const &ecb) {
-  if (auto *handle = dlsym(as_native_lib(lib), name)) {
-    return dynamic_function_symbol(
-        reinterpret_cast<c_function_pointer<void()>>(handle));
-  } else {
-    ecb(dynl_ec::symbol_not_found);
-    return nullptr;
-  }
+  return hh_find_function(lib, name, ecb);
 }
 
-dynamic_library_pointer *do_load_library(char const *p,
+dynamic_library_pointer *src_load_library(char const *p,
                                          error_callback const &ecb) {
-  if (auto *handle = dlopen(p, RTLD_NOW)) {
-    return as_placeholder(handle);
-  } else {
-    ecb(dynl_ec::library_not_found);
-    return nullptr;
-  }
+  return hh_load_library(p, ecb);
 }
 
-void do_release_library(dynamic_library_pointer *p) {
-  dlclose(as_native_lib(p));
-}
+void src_release_library(dynamic_library_pointer *p) { hh_release_library(p); }
 } // namespace dynl::_backend
